Improved public board design
