+++ components/index.ts
@@ -2,3 +2,4 @@
 export { ChecklistItem } from "./checklist-item";
 export { Note } from "./note";
 export { StickyNotesDemo } from "./sticky-notes-demo";
+export { BoardTopbar } from "./board-topbar";
